feat: 행사 설명 입력 필드 기본 정보란으로 이동
행사 기본 정보 입력에 설명 입력 추가, 레이아웃 변경사항 적용
